% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_config_server.R
\name{apiConfigServer}
\alias{apiConfigServer}
\title{API Configuration Server}
\usage{
apiConfigServer(id)
}
\arguments{
\item{id}{The namespace ID for the module.}
}
\description{
Server logic for the API Configuration UI.
}
