% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_config_ui.R
\name{apiConfigUI}
\alias{apiConfigUI}
\title{API Configuration UI}
\usage{
apiConfigUI(id)
}
\arguments{
\item{id}{The namespace ID for the module.}
}
\value{
A Shiny UI definition.
}
\description{
Creates a Shiny UI for configuring API providers.
}
