% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_shiny.R
\name{wrap_reactive_tools}
\alias{wrap_reactive_tools}
\title{Wrap Reactive Tools}
\usage{
wrap_reactive_tools(tools, rv, session)
}
\arguments{
\item{tools}{List of Tool objects, possibly including ReactiveTool objects.}

\item{rv}{The reactiveValues object to inject.}

\item{session}{The Shiny session object to inject.}
}
\value{
List of wrapped Tool objects ready for use.
}
\description{
Wraps reactive tools to inject reactiveValues and session into their
execute functions. Call this in your Shiny server before passing tools
to aiChatServer.
}
